test: colocate Python tests next to the code under test
Move the suite from the central ord_app/tests/ tree to <module>_test.py files
beside the code they exercise, and switch pytest to the *_test.py suffix.

- Endpoint/integration tests merge per feature into resources/v1/<feature>_test.py
  (datasets, reactions, group, templates, users, auth). The repository test moves
  to repositories/users_test.py.
- Shared fixtures/helpers move to ord_app/conftest.py (discovered package-wide);
  testdata stays under ord_app/tests/testdata/ (the path helper is unchanged).
- pytest config: testpaths = ["ord_app"], python_files = ["*_test.py"].
- ruff per-file-ignores retarget from ord_app/tests/** to **/*_test.py +
  **/conftest.py (the colocated files are no longer under tests/).
- Renamed one merge collision: the PATCH-on-missing-template test (formerly a
  second test_get_non_existent_template) is now test_update_non_existent_template.

Test bodies are unchanged; 74 test functions preserved, full suite green.
